+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 ## 2.0.0
 - Support for Polymer 2
 - Drop support for Node < 6 since Polymer 2 tools only support Node 6+
+- Replace ast-query with falafel for better support of ES2015 classes
 
 ## 1.1.0
 - Add wc-i18n v2 support